config: correct doc comment and add complex test
As a user unfamiliar with HCL, it was not clear how to define
an "object". I looked at the test and couldn't find anything testing
this functionality, and I also found inconsistent comments around
the use of regular expressions in the translation.

This commit adds a complex test example and corrects the comments.